+++ ledger/error/generator/BUILD.bazel
@@ -63,7 +63,8 @@ da_scala_library(
     tags = ["maven_coordinates=com.daml:error-generator-lib:__VERSION__"],
     visibility = ["//visibility:public"],
     runtime_deps = [
-        # Add the KV errors to the classpath so they can be picked up by the generator
+        # Add error definition targets to the classpath so they can be picked up by the generator
+        "//ledger/ledger-api-errors",
         "//ledger/participant-state-kv-errors",
     ],
     deps = [
