+++ ledger/ledger-api-auth/BUILD.bazel
@@ -29,6 +29,7 @@ da_scala_library(
         "//ledger/error",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-errors",
         "//ledger/participant-local-store",
         "//libs-scala/contextualized-logging",
         "@maven//:com_auth0_java_jwt",
@@ -77,6 +78,7 @@ da_scala_test_suite(
         "//ledger/error:error-test-lib",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-errors",
         "//ledger/participant-local-store",
         "//ledger/participant-state-index",
         "//ledger/test-common",
