+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/util/ObserverWithResult.scala
@@ -3,7 +3,7 @@
 
 package com.daml.ledger.api.benchtool.util
 
-import com.daml.error.definitions.LedgerApiErrors
+import com.daml.error.definitions.CommonErrors
 import com.daml.error.utils.ErrorDetails
 import io.grpc.stub.{ClientCallStreamObserver, ClientResponseObserver}
 import org.slf4j.Logger
@@ -48,7 +48,7 @@ abstract class ObserverWithResult[RespT, Result](logger: Logger)
   }
 
   private def isServerShuttingDownError(ex: io.grpc.StatusRuntimeException): Boolean =
-    ErrorDetails.matches(ex, LedgerApiErrors.ServerIsShuttingDown)
+    ErrorDetails.matches(ex, CommonErrors.ServerIsShuttingDown)
 
   override def onCompleted(): Unit = {
     logger.info(withStreamName(s"Stream completed."))
