+++ ledger/ledger-api-common/BUILD.bazel
@@ -39,6 +39,7 @@ da_scala_library(
         "//ledger/error",
         "//ledger/ledger-api-akka",
         "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-errors",
         "//ledger/ledger-api-health",
         "//ledger/ledger-offset",
         "//ledger/ledger-resources",
@@ -127,6 +128,7 @@ da_scala_test_suite(
         "//ledger/ledger-api-akka",
         "//ledger/ledger-api-client",
         "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-errors",
         "//ledger/ledger-api-health",
         "//ledger/ledger-offset",
         "//ledger/metrics",
