+++ ledger/ledger-api-errors/BUILD.bazel
@@ -1,0 +1,29 @@
+# Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_library",
+)
+
+da_scala_library(
+    name = "ledger-api-errors",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    tags = ["maven_coordinates=com.daml:ledger-api-errors:__VERSION__"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//daml-lf/archive:daml_lf_archive_reader",
+        "//daml-lf/data",
+        "//daml-lf/engine",
+        "//daml-lf/language",
+        "//daml-lf/transaction",
+        "//daml-lf/validation",
+        "//ledger-api/grpc-definitions:ledger_api_proto_scala",
+        "//ledger/error",
+        "//ledger/participant-integration-api:participant-integration-api-proto_scala",
+        "//ledger/participant-state",
+        "@maven//:com_google_api_grpc_proto_google_common_protos",
+        "@maven//:io_grpc_grpc_api",
+        "@maven//:org_slf4j_slf4j_api",
+    ],
+)
