+++ ledger/ledger-api-errors/src/main/scala/com/daml/error/definitions/LedgerApiErrors.scala
@@ -24,33 +24,14 @@ object LedgerApiErrors extends LedgerApiErrorGroup {
   val EarliestOffsetMetadataKey = "earliest_offset"
 
   @Explanation(
-    """This error category is used to signal that an unimplemented code-path has been triggered by a client or participant operator request."""
-  )
-  @Resolution(
-    """This error is caused by a participant node misconfiguration or by an implementation bug.
-      |Resolution requires participant operator intervention."""
-  )
-  object UnsupportedOperation
-      extends ErrorCode(
-        id = "UNSUPPORTED_OPERATION",
-        ErrorCategory.InternalUnsupportedOperation,
-      ) {
-
-    case class Reject(message: String)(implicit errorLogger: ContextualizedErrorLogger)
-        extends DamlErrorWithDefiniteAnswer(
-          cause = s"The request exercised an unsupported operation: $message"
-        )
-  }
-
-  @Explanation(
     """This error occurs when a participant rejects a command due to excessive load.
-        |Load can be caused by the following factors:
-        |1. when commands are submitted to the participant through its Ledger API,
-        |2. when the participant receives requests from other participants through a connected domain."""
+      |Load can be caused by the following factors:
+      |1. when commands are submitted to the participant through its Ledger API,
+      |2. when the participant receives requests from other participants through a connected domain."""
   )
   @Resolution(
     """Wait a bit and retry, preferably with some backoff factor.
-        |If possible, ask other participants to send fewer requests; the domain operator can enforce this by imposing a rate limit."""
+      |If possible, ask other participants to send fewer requests; the domain operator can enforce this by imposing a rate limit."""
   )
   object ParticipantBackpressure
       extends ErrorCode(
@@ -135,7 +116,7 @@ object LedgerApiErrors extends LedgerApiErrorGroup {
       |1. Review the historical 'queue size' growth by inspecting the metric given in the message.
       |2. Review the maximum 'queue size' limits configured in the rate limiting configuration.
       |3. Try to space out requests that are likely to require a lot of CPU or database power.
-      """
+     """
   )
   object ThreadpoolOverloaded
       extends ErrorCode(
@@ -158,56 +139,6 @@ object LedgerApiErrors extends LedgerApiErrorGroup {
             "metricPrefix" -> metricPrefix,
             "fullMethodName" -> fullMethodName,
           ),
-        )
-  }
-
-  @Explanation(
-    "This rejection is given when a request processing status is not known and a time-out is reached."
-  )
-  @Resolution(
-    "Retry for transient problems. If non-transient contact the operator as the time-out limit might be too short."
-  )
-  object RequestTimeOut
-      extends ErrorCode(
-        id = "REQUEST_TIME_OUT",
-        ErrorCategory.DeadlineExceededRequestStateUnknown,
-      ) {
-    case class Reject(message: String, override val definiteAnswer: Boolean)(implicit
-        loggingContext: ContextualizedErrorLogger
-    ) extends DamlErrorWithDefiniteAnswer(
-          cause = message,
-          definiteAnswer = definiteAnswer,
-        )
-  }
-
-  @Explanation("This rejection is given when the requested service has already been closed.")
-  @Resolution(
-    "Retry re-submitting the request. If the error persists, contact the participant operator."
-  )
-  object ServiceNotRunning
-      extends ErrorCode(
-        id = "SERVICE_NOT_RUNNING",
-        ErrorCategory.TransientServerFailure,
-      ) {
-    case class Reject(serviceName: String)(implicit
-        loggingContext: ContextualizedErrorLogger
-    ) extends DamlErrorWithDefiniteAnswer(
-          cause = s"$serviceName has been shut down.",
-          extraContext = Map("service_name" -> serviceName),
-        )
-  }
-
-  @Explanation("This rejection is given when the participant server is shutting down.")
-  @Resolution("Contact the participant operator.")
-  object ServerIsShuttingDown
-      extends ErrorCode(
-        id = "SERVER_IS_SHUTTING_DOWN",
-        ErrorCategory.TransientServerFailure,
-      ) {
-    case class Reject()(implicit
-        loggingContext: ContextualizedErrorLogger
-    ) extends DamlErrorWithDefiniteAnswer(
-          cause = "Server is shutting down"
         )
   }
 
@@ -283,5 +214,4 @@ object LedgerApiErrors extends LedgerApiErrorGroup {
         loggingContext: ContextualizedErrorLogger
     ) extends DamlErrorWithDefiniteAnswer(cause = message, throwableO = throwableO)
   }
-
 }
