+++ ledger/ledger-api-errors/src/main/scala/com/daml/error/definitions/groups/AdminServices.scala
@@ -3,8 +3,9 @@
 
 package com.daml.error.definitions.groups
 
+import com.daml.error.definitions.{DamlErrorWithDefiniteAnswer, LedgerApiErrors}
+import com.daml.error.definitions.groups
 import com.daml.error.{ContextualizedErrorLogger, ErrorCategory, ErrorCode, Explanation, Resolution}
-import com.daml.error.definitions.{DamlErrorWithDefiniteAnswer, LedgerApiErrors, groups}
 
 @Explanation("Errors raised by Ledger API admin services.")
 object AdminServices extends LedgerApiErrors.AdminServicesErrorGroup {
