+++ ledger/ledger-api-errors/src/main/scala/com/daml/error/definitions/groups/WriteServiceRejections.scala
