+++ ledger/ledger-api-tests/infrastructure/BUILD.bazel
@@ -25,6 +25,7 @@ load(
             "//ledger-api/grpc-definitions:ledger_api_proto_scala",
             "//ledger/error",
             "//ledger/ledger-api-common",
+            "//ledger/ledger-api-errors",
             "//libs-scala/test-evidence/tag:test-evidence-tag",
             "//ledger/test-common:dar-files-%s-lib" % lf_version,
             "//ledger/test-common:test-common-%s" % lf_version,
