+++ ledger/ledger-api-tests/suites/deps.bzl
@@ -16,6 +16,7 @@ def deps(lf_version):
         "//daml-lf/transaction",
         "//language-support/scala/bindings",
         "//ledger/error",
+        "//ledger/ledger-api-errors",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-tests/infrastructure:infrastructure-%s" % lf_version,
         "//ledger/ledger-resources",
