+++ ledger/participant-integration-api/BUILD.bazel
@@ -38,11 +38,13 @@ compile_deps = [
     "//daml-lf/language",
     "//daml-lf/transaction",
     "//daml-lf/transaction:value_proto_java",
+    "//daml-lf/validation",
     "//language-support/scala/bindings",
     "//ledger-api/rs-grpc-akka",
     "//ledger-api/rs-grpc-bridge",
     "//ledger/caching",
-    "//ledger/error:error",
+    "//ledger/error",
+    "//ledger/ledger-api-errors",
     "//ledger/ledger-api-akka",
     "//ledger/ledger-api-auth",
     "//ledger/ledger-api-client",
@@ -287,6 +289,7 @@ da_scala_test_suite(
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-common:ledger-api-common-scala-tests-lib",
         "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-errors",
         "//ledger/ledger-api-health",
         "//ledger/ledger-configuration",
         "//ledger/ledger-offset",
