+++ ledger/participant-integration-api/src/main/scala/platform/DispatcherState.scala
@@ -4,7 +4,7 @@
 package com.daml.platform
 
 import com.daml.error.DamlContextualizedErrorLogger
-import com.daml.error.definitions.LedgerApiErrors
+import com.daml.error.definitions.CommonErrors
 import com.daml.ledger.offset.Offset
 import com.daml.ledger.resources.ResourceOwner
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
@@ -127,7 +127,7 @@ class DispatcherState(dispatcherShutdownTimeout: Duration)(implicit
       loggingContext = loggingContext,
       None,
     )
-    LedgerApiErrors.ServiceNotRunning.Reject(ServiceName)(contextualizedErrorLogger).asGrpcError
+    CommonErrors.ServiceNotRunning.Reject(ServiceName)(contextualizedErrorLogger).asGrpcError
   }
 }
