+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/services/tracking/QueueBackedTracker.scala
@@ -8,7 +8,7 @@ import akka.stream.{BoundedSourceQueue, Materializer, QueueOfferResult}
 import akka.{Done, NotUsed}
 import com.daml.metrics.api.MetricHandle.{Counter, Timer}
 import com.daml.error.DamlContextualizedErrorLogger
-import com.daml.error.definitions.LedgerApiErrors
+import com.daml.error.definitions.{CommonErrors, LedgerApiErrors}
 import com.daml.ledger.client.services.commands.CommandSubmission
 import com.daml.ledger.client.services.commands.CommandTrackerFlow.Materialized
 import com.daml.ledger.client.services.commands.tracker.CompletionResponse._
@@ -67,7 +67,7 @@ private[services] final class QueueBackedTracker(
         )
       case Success(QueueOfferResult.QueueClosed) =>
         toQueueSubmitFailure(
-          LedgerApiErrors.ServiceNotRunning
+          CommonErrors.ServiceNotRunning
             .Reject("Command service queue")
             .asGrpcStatus
         )
