+++ ledger/participant-integration-api/src/main/scala/platform/store/dao/events/LfValueTranslation.scala
@@ -4,7 +4,6 @@
 package com.daml.platform.store.dao.events
 
 import com.daml.error.DamlContextualizedErrorLogger
-import com.daml.error.definitions.{ErrorCause, RejectionGenerators}
 
 import java.io.ByteArrayInputStream
 import com.daml.ledger.api.v1.event.{CreatedEvent, ExercisedEvent, InterfaceView}
@@ -22,6 +21,7 @@ import com.daml.lf.value.Value.VersionedValue
 import com.daml.lf.{engine => LfEngine}
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.metrics.{Metrics, Timed}
+import com.daml.platform.apiserver.services.{ErrorCause, RejectionGenerators}
 import com.daml.platform.{
   ContractId,
   Create,
