+++ ledger/sandbox-on-x/src/main/scala/com/daml/ledger/sandbox/domain/Rejection.scala
@@ -7,12 +7,13 @@ package ledger.sandbox.domain
 import com.daml.ledger.participant.state.v2.{ChangeId, CompletionInfo, Update}
 import com.daml.ledger.participant.state.v2.Update.CommandRejected.FinalReason
 import error.ContextualizedErrorLogger
-import error.definitions.LedgerApiErrors
+import error.definitions.{CommonErrors, LedgerApiErrors}
 import ledger.configuration.LedgerTimeModel
 import com.daml.lf.data.Time.Timestamp
 import com.daml.lf.transaction.GlobalKey
 import com.daml.lf.value.Value.ContractId
 import com.google.rpc.status.Status
+
 import java.time.Duration
 
 private[sandbox] sealed trait Rejection extends Product with Serializable {
@@ -64,7 +65,7 @@ private[sandbox] object Rejection {
   final case class OffsetDeduplicationPeriodUnsupported(completionInfo: CompletionInfo)(implicit
       contextualizedErrorLogger: ContextualizedErrorLogger
   ) extends Rejection {
-    override def toStatus: Status = LedgerApiErrors.UnsupportedOperation
+    override def toStatus: Status = CommonErrors.UnsupportedOperation
       .Reject("command deduplication with periods specified using offsets")
       .rpcStatus()
   }
