+++ release/artifacts.yaml
@@ -277,3 +277,5 @@
   type: jar-scala
 - target: //ledger/error/generator:lib
   type: jar-scala
+- target: //ledger/ledger-api-errors:ledger-api-errors
+  type: jar-scala
